bin/orgmk-init: Allow users to provide flags in the call of emacs
This commit introduces a mechanism for allowing users to provide flags
to the call of $EMACS_BATCH in bin/orgmk-init.  With this change, it
is possible to do calls like this:

    ORGMK_FLAGS="-u myusername" org2pdf input.org

The exemple above is quite handy, because the --batch option, which is
used in the call of emacs, prevents the personal initialization file
to be loaded.

The unused varible ORG_FLAGS in orgmk-init is changed into
ORGMK_FLAGS, because this is a more explicit name.  This is only a
sugegstion and perhaps another better solution could be provided.  The
essence of this commit is to allow an environment variable (in
occurrence, ORGMK_FLAGS) to sneak into orgmk-init.